Fix lighten and shade mode
Program now works mostly fine.
Removed all the Mode() functions and incoroporated the different
modes into the boxListen() function that will color the box in
different ways based on boolean values.
Changed base black color to hex value #101010 because attempting to
lighten #000000 doesn't seem to work.

Note:
Some rainbow colors seem to be in the format of rgba instead of
rgb. This causes them to be unaffected by lightening and shading.